Rename generated top-level directories.
I find the long-standing attachment (purely for historical reasons, as far as I
can tell) to directory names that seem to have arbitrary contractions and
pluralization utterly baffling.  While I agree that consistency and patterns are
important to reduce friction, the tooling available makes the impact of
improvment virtually zero.

The changes applied here create the following top-level project directory
layout:

 - resource
 - source
 - test

The contraction of `src` is unecessarily inconsistent with both `test` and
`resources`.  The pluralization of `resources` is itself again arbitrarily
inconsistent with `source`.