Prefer `:require :refer` over `:require :refer :all`.
See: https://github.com/bbatsov/clojure-style-guide#prefer-require-over-use

Since `clojure.test` is the ns that is depended upon for functionality, it is
reasonable to bring in only what it needed (as we would for any other namespace
where only a small subset of the required functionality is routinely used).  On
the other hand, the main ns is the subject under test (i.e. the thing being
analyzed), and contextually therefore it makes sense that all of its features
should be available.