Strip generated project README to what is minimally necessary.
FIXME's and TODO's are a useful way of calling attention to items that need to
be addressed, guiding the way through setting up a project.  If one is not ready
to do that work, however, it leaves the project looking ... unkempt.  _In my
opinion_, this is an instance where it is better to have nothing than something
that appears to be half-done.

The generated README includes the project licencing terms, so - for safety - in
the case of accidental publication, the copyright is announced with all rights
reserved, and an open licence can be adopted at the discretion of the developer.
The generated project LICENSE file has been removed for the same reason.