Remove CHANGELOG and documentation placeholders.
FIXME's and TODO's are a useful way of calling attention to items that need to
be addressed, guiding the way through setting up a project.  If one is not ready
to do that work, however, it leaves the project looking ... unkempt.  _In my
opinion_, this is an instance where it is better to have nothing than something
that appears to be half-done.

Related: 593966a0c